test(injected-deps-syncer): move the patcher-driven cases to DirPatcher.test.ts
Three of the cases covering a skipped inode drive DirPatcher.apply() rather
than extendFilesMap, so they belong beside the other patcher tests, where
createDir and createFile already exist. A createFifo helper joins them.

extendFilesMap.test.ts keeps the two that assert the shape of the inode map.
